move tests of python API that require erlang into a designated integration test file.
